feat: added weitere klassifikationen mapper (#296)
* feat: 295 with custom mapping config

* feat: added weitere klassifikationen mapper

* fix: if histologie_id is null, fall back to Meldung ID

* fix missing class

* rm typo'd snapshots

* fix ordinal value extension type

* fix strahlentherapie url

* fix tnm rebase

[skip ci]

+++ src/main/java/org/miracum/streams/ume/obdstofhir/mapper/mii/GleasonScoreMapper.java
@@ -174,7 +174,7 @@ public class GleasonScoreMapper extends ObdsToFhirMapper {
 
     var gleasonScoreNumeric = matcher.group(1);
     scoreCoding.addExtension(
-        fhirProperties.getExtensions().getOrdinalValue(), new IntegerType(gleasonScoreNumeric));
+        fhirProperties.getExtensions().getOrdinalValue(), new DecimalType(gleasonScoreNumeric));
 
     var concept = new CodeableConcept(scoreCoding).setText(scoreErgebnis);
     observation.setValue(concept);
